requirements.txt
boto==2.39.0
defusedxml==0.4.1
dj-database-url==0.4.0
Django==1.8.2
django-haystack==2.4.1
django-storages-redux==1.3.2
easy-thumbnails==2.2
elasticsearch==2.1.0
gunicorn==19.4.5
oauthlib==1.0.3
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.0
python-social-auth==0.2.14
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
sorl-thumbnail==12.3
urllib3==1.12
